refactor: split plugin file into targets
