+++ projects/build-env/src/executors/bootstrap/constants.ts
@@ -1,1 +1,8 @@
+import {join} from "node:path";
+
 export const VERDACCIO_REGISTRY_JSON = 'verdaccio-registry.json';
+export const DEFAULT_VERDACCIO_STORAGE_DIR = join(
+  'tmp',
+  'local-registry',
+  'storage'
+);
