+++ projects/build-env/src/executors/bootstrap/verdaccio-registry.ts
@@ -5,9 +5,9 @@ import { executeProcess } from '../../internal/execute-process';
 import { uniquePort } from './unique-port';
 import { formatError, formatInfo } from '../../internal/logging';
 import {
-  DEFAULT_START_VERDACCIO_TARGET,
-  DEFAULT_VERDACCIO_STORAGE_DIR,
-} from '../../internal/constants';
+  DEFAULT_START_VERDACCIO_TARGET
+} from '../../plugin/targets/environment.targets';
+import {DEFAULT_VERDACCIO_STORAGE_DIR} from "./constants";
 
 const VERDACCIO_TOKEN = 'Verdaccio: ';
