+++ projects/build-env/src/executors/npm-install/executor.unit-test.ts
@@ -4,8 +4,7 @@ import { MEMFS_VOLUME } from '@push-based/test-utils';
 import * as execProcessModule from '../../internal/execute-process';
 import { logger, readJsonFile } from '@nx/devkit';
 import { join } from 'node:path';
-import { DEFAULT_ENVIRONMENTS_OUTPUT_DIR } from '../../internal/constants';
-
+import { DEFAULT_ENVIRONMENTS_OUTPUT_DIR } from '../../plugin/constants';
 vi.mock('@nx/devkit', async () => {
   const actual = await vi.importActual('@nx/devkit');
   return {
