+++ projects/build-env/src/executors/npm-publish/constants.ts
@@ -1,0 +1,2 @@
+export const NPMRC_FILENAME = '.npmrc';
+
