+++ projects/build-env/src/plugin/constants.ts
@@ -1,0 +1,4 @@
+import {join} from "node:path";
+
+export const DEFAULT_OPTION_ENVIRONMENT_TARGET_NAMES = ['e2e'];
+export const DEFAULT_ENVIRONMENTS_OUTPUT_DIR = join('tmp', 'environments');
